back/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
datasets==4.0.0
deprecation==2.1.0
dill==0.3.8
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
postgrest==1.1.1
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
